// app/features/quiz/generated/models/knowde.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { Additional } from "./additional";
import type { KStats } from "./kStats";
import type { Term } from "./term";

/**
 * 知識の最小単位.
 */
export interface Knowde {
  sentence: string;
  uid: string;
  term?: Term | null;
  additional?: Additional | null;
  stats: KStats;
  resource_uid: string;
}