// app/features/quiz/generated/models/quizChainRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

/**
 * Quizから見たSentenceの役割.
 */
export type QuizChainRole = (typeof QuizChainRole)[keyof typeof QuizChainRole];

export const QuizChainRole = {
  target: "target",
  option: "option",
  correct: "correct",
} as const;